many changes

+++ app.py
@@ -3,6 +3,7 @@ from flask import Flask, request, abort, jsonify
 from flask_sqlalchemy import SQLAlchemy
 from flask_cors import CORS
 from models import setup_db, Movie, Actor
+from auth import AuthError, requires_auth
 
 def create_app(test_config=None):
     # create and configure the app
@@ -15,7 +16,8 @@ def create_app(test_config=None):
         return 'Hello_world'
 
     @app.route('/actors', methods=["POST"])
-    def post_actor():
+    @requires_auth('post:actor')
+    def post_actor(token):
         body = request.get_json()
         name = body.get('name', None)
         age = body.get('age', None)
@@ -32,8 +34,11 @@ def create_app(test_config=None):
             abort(422)
 
     @app.route('/actors/<int:actor_id>', methods=["DELETE"])
-    def delete_actor(actor_id):
+    @requires_auth('delete:actor')
+    def delete_actor(token, actor_id):
         actor = Actor.query.filter(Actor.id == actor_id).one_or_none()
+        if actor is None:
+            abort(404)
         try:
             actor.delete()
             all_actors = [actor.format() for actor in Actor.query.all()]
@@ -45,7 +50,8 @@ def create_app(test_config=None):
             abort(422)
 
     @app.route('/movies', methods=["POST"])
-    def post_movie():
+    @requires_auth('post:movie')
+    def post_movie(token):
         body = request.get_json()
         title = body.get('title', None)
         release_date = body.get('release_date', None)
@@ -61,8 +67,11 @@ def create_app(test_config=None):
             abort(422)
 
     @app.route('/movies/<int:movie_id>', methods=["DELETE"])
-    def delete_movie(movie_id):
+    @requires_auth('delete:movie')
+    def delete_movie(token, movie_id):
         movie = Movie.query.filter(Movie.id == movie_id).one_or_none()
+        if movie is None:
+            abort(404)
         try:
             movie.delete()
             all_movies = [movie.format() for movie in Movie.query.all()]
@@ -74,23 +83,30 @@ def create_app(test_config=None):
             abort(422)
 
     @app.route('/movies',methods=["GET"])
-    def get_movies():
+    @requires_auth('get:movies')
+    def get_movies(token):
         all_movies = [movie.format() for movie in Movie.query.all()]
+        if len(all_movies) == 0:
+            abort(422)
         return jsonify({
             "success": True,
             "movies": all_movies
         })
 
     @app.route('/actors',methods=["GET"])
-    def get_actors():
+    @requires_auth('get:actors')
+    def get_actors(token):
         all_actors = [actor.format() for actor in Actor.query.all()]
+        if len(all_actors) == 0:
+            abort(422)
         return jsonify({
             "success": True,
             "actors": all_actors
         })
 
     @app.route('/actors/<int:actor_id>', methods=["PATCH"])
-    def patch_actor(actor_id):
+    @requires_auth('patch:actor')
+    def patch_actor(token, actor_id):
         actor = Actor.query.filter(Actor.id == actor_id).one_or_none()
         if actor is None:
             abort(404)
@@ -109,7 +125,8 @@ def create_app(test_config=None):
             abort(422)
 
     @app.route('/movies/<int:movie_id>', methods=["PATCH"])
-    def edit_movie(movie_id):
+    @requires_auth('patch:movie')
+    def patch_movie(token, movie_id):
         movie = Movie.query.filter(Movie.id == movie_id).one_or_none()
         if movie is None:
             abort(404)
@@ -139,21 +156,21 @@ def create_app(test_config=None):
         return jsonify({
             "success": False,
             'error': 422,
-            'message': 'unprocessible'})
-
-    @app.errorhandler(500)
-    def server_error(error):
-        return jsonify({
-            "success": False,
-            'error': 500,
-            'message': 'internal server error'})
+            'message': 'unprocessible'}),422
 
     @app.errorhandler(401)
     def unauthorized(error):
         return jsonify({
             "success": False,
             'error': 401,
-            'message': 'unauthorized'})
+            'message': 'unauthorized'}),401
+
+    @app.errorhandler(AuthError)
+    def handle_auth_error(ex):
+        response = jsonify(ex.error)
+        response.status_code = ex.status_code
+        return response
+
     return app
 
 app = create_app()
